Switch object-store metadata to use utils crate
The config struct for metadata client has switched from an explicit bucket name
to a URL and renamed to "path", consistent with snapshots configuration.

Additionally we now use the shared object store options struct which allows
users to directly set a variety of credentials config keys.

Finally, we switch to using our object-store-utils crate which handles the
client building and AWS credentials caching. This allows us to reuse the AWS
SDK conventions for more ergonomic and consistent configuration.