Only trim multi-node clusters when Archived LSN is reported
This change makes it so that in Restate deployments with more than one node,
a snapshot repository is a requirement for auto-trimming of logs. Previously
we would have trimmed logs based on the reported persisted LSN, which isn't
always safe considering that additional nodes may be added at a later time.