[sql] Reduce the number of in-flight RecordBatchs (#2841)
This commit sets the number of record batches that has to be kept in
memory during a scan (per-partition) to 2.
With this we ensure that at most 2048 rows are materialized at a time per
scanner. One is being filled during a scan, and an other is pending to be
processed by the parent operator (the consumer of the scan)