Flush config file to disk before starting Restate process
Dropping the file alone, does not guarantee that the file content is
written to disk and the file being closed immediately. If the config
file is not written when starting the Restate process, then it will
start with the default configuration which causes the raft_metadata_cluster_chaos_test
to fail every now and then.

This fixes #2828.