Don't panic if we cannot obtain connection attemp result
For logging purposes we try to obtain the result of the connection attempt
in a synchronous function if the connection attempt task has finished.
Unfortunately, we cannot rely on TaskHandle::is_finished for task_handle.now_or_never()
to always return Some(...). The problem is that the underlying Tokio JoinHandle participates
in the cooperative task scheduling and if we ran out of budget, then it would return
Poll::Pending. That's why we ignore the result in this case and continue.

This fixes #2879.