[Core] Remove WeakConnection
Removes WeakConnection as it's not needed after having tighter control over the egress sender's lifecycle. This is a critical step towards simpler interfaces. It also gets us closer to the ability of sending responses/messages during drain. The behaviour of the reactor and drain will change in the next set of PRs.

- Removes extra indirections, upgrades, etc.
- Minor changes to the MessageRouter, more to come.
```
// intentionally empty
```