Use stable network identifier when starting a K8s based Restate cluster
W/o stable network identifier, it is not possible that the replicated metadata
cluster survives a crash of multiple pods (N/2 + 1) because they would be restarted
with a different IP which makes them no longer reachable by the other nodes
(based on the advertised address stored before in the NodesConfiguration). As a
result, the metadata cluster would lose write availability because it can no longer
form a majority quorum.

To solve this problem, we are now using the stable network identifier provided by
the stateful set. It requires to run a headless service with the same name as the
one configured as the serviceName in the stateful set spec. Now, all Restate pods
are started with this stable network identifier as their advertised address.
Additionally, we set the podManagementPolicy to parallel to enable that a metadata
cluster majority quorum can form in case of a multiple pod failures. Moreover, we
configure the headless service to publish the pod addresses before they are ready.