Fail read-only requests that get silently dropped by Raft
During a leadership change a new leader might not have fully re-established
what's the latest commit index is. To do this, a new leader upon winning the
leadership writes an empty entry. In order to avoid stale reads during this
time, read-index requests are silently ignored by RawNode. If this is not
correctly handled, it can lead to unanswered metadata requests. To solve
the problem, we are comparing the previous pending reads and ready reads
with after calling RawNode::read_index. If they are still the same, then we
faill the request with RequestError::Unavailable (as if we aren't the leader
yet). It's then the responsibility of the client to retry the operation.

This fixes #2947.