Only update metadata manager if metadata was actually written to the metadata store
The Scheduler updates the PartitionTable whenever it updates the partition placement.
Once it decides on a new PartitionTable, it tries to store it in the metadata store.
If the write fails because of a precondition violation, we were accidentally publishing
the rejected PartitionTable. This can lead to cluster nodes operating on different partition
tables. One symptom of it is that an ingress becomes unresponsive because the CC selected a
different leader than what the ingress believes is the current leader. Therefore, the ingress
sends the ingress invocation to the wrong partition processor where it gets rejected.

Additionally, to speed up convergence when detecting a concurrent write to the partition table,
we are trying to sync the metadata based on the expected version + 1 instead of the new version
+ 1 which might not have been written at this point.

This commit also includes a few logging improvements which helped while debugging the above-described
problem.

This fixes #2959.