Default bind port to 5122 when inferring from ambiguous advertised-address (#2941)
This change updates the implied bind-address logic to default to 5122 if only a hostname is
specified, without a protocol or port.

In cases where a protocol is specified, the specific protocol determines the actual bind port - we
print a startup warning when this happens as this might be surprising. HTTPS is an exception since
restate-server does not natively support this; it's not an error to advertise an https:// address
but it is a misconfiguration to not specify a bind-address in that case.

Fixes #2940