[Bifrost] Front-loading serde work for sealed chains
This front-loads necessary work to allow future versions to set a chain to be permanently sealed. This is currently not used and in v1.3, restate will respect if a chain is sealed but it'll not attempt to seal a chain. This work is intended to ensure v1.3 doesn't clobber this field if written by future versions.