Add support for removing/adding metadata server nodes
This commit adds restatectl metadata-server add-node/remove-node to add and remove
nodes to and from the metadata cluster. Internally, an add command instructs the
joining node to reach out to the leader to ask for a reconfiguration. A remove command
is sent to the leader where it triggers a reconfiguration to remove the given node. To
also stop removed metadata cluster members that have missed the configuration change,
we are using the MetadataServerState stored in the NodesConfiguration as an additional
signal. If a node detects that its MetadataServerState is set to Standby while it is
running as a Member, then it will step down. Likewise, if a Standby node detects that its
MetadataServerState is Member, then it will try to join the cluster because it might mean
that it tried joining before but failed just before receiving the response.

This fixes #2994.