[Storage] Forward compatibility with other codecs
Adds support for decoding bincode 2.0 (it's finally released) and json. Encoding is still with the normal flexbuffers but this enables v1.3 to decode those in the future.
Also removed some unnecessary checks that were meaningful when Encode/Decode traits were not dyn-compatible. Right now, we only accept BytesMut.

For all codec types, deserialization requires contiguous buffers which is already the case for rocksdb slices and our network message decoding. In the future, the API will be updated to reflect this.

```
// intentionally empty
```