[rocksdb] Disable Direct I/O by default
This is now disabled by default, but it's still advised to enable it on non-network filesystems with reliable device controllers. The reason for this is that we have observed a bug when running RocksDB+Direct_IO on EBS volumes. The bug results in multi-get operations failing with a "corruption" error on SST files.