[LogServer] Enable parallel compactions
This enables sub-compactions (aka. parallel compactions) in log-server to leverage the CPU capacity to speed up compactions. Prior to this, we were limiting compactions to 1 background this, with this change, compactions (for log-server) will use half the number of cores by default. This also adds a configuration option to customize it.

The same can be considered for partition store in the future but we need to be careful as rocksdb will not limit parallelism across databases, so it's easy to go beyond all available background threads.

```
// intentionally empty
```