[Core] Reduce default concurrent connections to 3
As of v1.3, every node will maintain a set of "outgoing/forward" connections to other peers, this means that effectively in bidirectional setup, we'll 2x this value in total connections. Reducing the default value from 13 to 3 in this setup didn't result in performance impact in load tests, therefore, it's better to reduce it.

```
// intentionally empty
```