[Bifrost] auto-seal sequencers on graceful shutdown
This PR is another step towards better graceful shutdown. We'll ask Bifrost's watchdog to shutdown before we terminate connections to give it a chance to cleanly seal all local sequencers. This also fixes a bug where the rocksdb writer might get terminated before the loglet workers, causing the loglet workers to spew errors on shutdown and not cleanly draining up.

```
// intentionally empty
```