[LogServer] Add `rocksdb-max-wal-size` to `[log-server]` to limit WAL size
This adds a new option `rocksdb-max-wal-size` in `[log-server]` with hidden default (0) that controls the maximum size of the WAL in bytes. If the WAL size exceeds this limit, the server will force a flush to the SST files. This is useful to prevent the WAL from growing indefinitely and consuming disk space.
The default (0) will use twice the total size allocated to memtables for log-server by default, I've also adjusted the number of WAL files to recycle between to 2 to match this value.

With this change, there is no need for atomic_flushes anymore, so I've removed it.


Manually tested, old WAL files are removed as expected.
```
// intentionally empty
```