Limit scope of local-loglet feature to bifrost crate
This commit simplifies how we use the local-loglet feature by limiting it to the
Bifrost crate only. This has the advantage that restate-types can use ProviderKind::Local
as the default value for the Bifrost provider. Moreover, it reduces the number of features
that might be misconfigured which usually shows up at release time.

This fixes #3118.