Set default values for deprecated options
With this commit we are now properly setting the default values when using a
deprecated option. This is particularly relevant if the deprecated option is a
struct that contains mutliple values and the user only specifies a few of them.
In this case, we want to a) not fail because not all fields have been specified
and b) fill in the default values that are specified in the configuration.
Before, the problem was that default values weren't respected for flattened structs
(see https://github.com/serde-rs/serde/issues/1879). Moreover, if a field of a sub-
struct was specified, then the default value of the sub-struct would be used instead
of the default value specified in the parent. This was problematic with RocksDbOptions,
for example. Since all fields are optional, it would deserialize as all not specified
fields being None. If the parent enabled the WAL, then it would not have been respected.

This fixes #3139.