Don't enforce DBRecoveryMode::AbsoluteConsistency for the metadata server
Since we cannot prevent tail writes from being incomplete in the presence of
hard crashes, using the DBRecoveryMode::AbsoluteConsistency might be too
restrictive because it prevents the node from restarting. If we haven't completed
the tail write, then we can be sure that we haven't acted on it yet. Therefore,
it is ok to resume and ignore it. The only problem is that the tail entry might
also have been corrupted after having been acted upon. This commit weakens the
recovery mode in favor of DBRecoveryMode::TolerateCorruptedTailRecords.

This fixes #3154.