[Bifrost] Improves appender behaviour when detecting seals
When a seal is detected, the appender will be smart about monitoring metadata version changes. It'll also not advance the sleep timer unless it spent the full sleep time waiting for metadata update. This improves its reaction time to reconfiguration changes. Primarily, it'll not wait for the full sleep duration if new log chain was detected.

```
// intentionally empty
```