[Core] New global metadata synchronization infrastructure
The new global metadata synchronization infrastructure is designed to
ensure that all nodes in the cluster have consistent metadata while reducing load over the metadata store.

This has significant implications for the overall performance and reliability of the system, especially when metadata store is overloaded or naturally slow.
The new infrastructure removes the need for adhoc metadata sync requests and limits the staleness to a configurable value.
Additionally, the new infrastructure will allow nodes to keep clawing metadata changes from the growing number of peers that report the new metadata version to it.

Other minor changes:
- Removed the option to manually `sync` metadata when requesting to create a snapshot
- On node startup, we wait for a full global metadata sync
- Added initially loaded metadata versions to the `My Node ID` log message


Note that there can be certain scenarios when an isolated node doesn't hear about a metadata change that happened in the cluster (a node that normally doesn't interact with the rest of the cluster). This is still covered by the metadata update interval which will eventually fetch the metadata from metadata store, and in the future, this will be automatically covered by gossip messaging disseminating metadata version changes much quicker than this.

```
// intentionally empty
```