[Core] Introducing network swimlanes
Network swimlanes categorize network connections into different groups based on their characteristics. This PR is a continuation of the work done in the previous PR, which introduced the concept of swimlanes. The goal is to enhance the network management capabilities of the system by allowing for more granular control and monitoring of network connections.

Most notable changes:
- We keep a single connection per swimlane instead of a pool of connections. This removes the configuration option `num-concurrent-connections`.
- Connections's window sizes are now configurable and get different defaults for data vs. control lanes
- Removal of `outbound-queue-length` in networking in favor of relying on memory-level settings (stream window sizes)
- Introducing `swimlane` in `Hello` message to communicate the swimlane to the peer during handshake