Efficiently pack metadata versions in network structs
Leveraging the fact the Version::INVALID is 0 to pack the metadata versions better in memory. Version needs 4 bytes (u32) while Option<Version> needs 8. `PeerMetadataVersion` now needs 16 bytes instead of 32, same for `Header`.

The other change is to let `MetadataVersions` keep track of the actual observed version of the peer as this might come handy for introspection later. This doesn't impact performance as we still only notify if the version is higher than our latest local view.