Introduce ReadIsolation for Storage transactions
This commit introduces the ReadIsolation level for Storage transactions.
This allows to choose between no or snapshot read isolation. The latter
is useful when reading from outside of the partition processor where
concurrent writes can happen (e.g. the invoker reading the journal, the
metadata and the state of an invocation). The snapshot isolation level
creates a Rocksdb snapshot which will affect compaction. It is,
therefore, recommended to not keep snapshot transactions open for too long.

This fixes #3192.