Leave member state when removing yourself from Raft configuration
Instead of waiting only for the NodesConfiguration to tell us when we switched from
Member to Standby state, this commit directly reacts to a configuration change where
we removed ourselves. Otherwise, we might still process responses from other nodes
because updating of the NodesConfiguration in the Metadata happens asynchronously from
the KvMemoryStorage. Processing messages after we have removed ourselves can result in
panics from the Raft library which doesn't expect this to happen.

This fixes #3204.