[Core] WireEncode accepts a reference instead of an owned value
This is necessary to allow `Box<dyn WireEncode>` to be used in the next PR. This also aligns with how most serialization libraries work, there is no need to send the owned value. I've temporarily disabled bilrost's IntoDto until we can get it to work with this design. Additionally. I didn't change the API of the send() API in network code to avoid making this as a sweeping change. If the need for passing the message by reference on that layer rises, we can change it.

```
// intentionally empty
```