introduce BilrostAs derive macro
Summary:
The macro can be deriven on anytime given that the target type
can be created From<&T>.

There is also an implementation for an adaptor that can use the
`Display` and `FromStr` to serialize and deserialize the type.

## Notice:
Right now the BilrostAs macro requires that the annotated type must
implement `Default`. This is needed since Bilrost requires that it
can create an `empty state` of that field.