Removes memory-loglet gating except for bifrost
The strategy here is to enable in-memory loglet in restate-types to ensure we don't fail deserialization or handling of all provider kinds in all parts of the code. The `in-memory` variant is hidden from command line options and from configuration documentation. However, memory-loglet is still gated in bifrost itself which means that the in-memory loglet implementation will not be running by default.

Bonus:
- Removes unnecessary dependencies from restatectl
- Removes stale `restatectl metadata` options that were not cleaned up after moving to metadata proxy