[restatectl] Remove dependency on RocksDB
Re. memory-loglet:
The strategy here is to enable in-memory loglet in restate-types to ensure we don't fail deserialization or handling of all provider kinds in all parts of the code. The `in-memory` variant is hidden from command line options and from configuration documentation. However, memory-loglet is still gated in bifrost itself which means that the in-memory loglet implementation will not be running by default.

Summary of changes:
- Introduces separate crate for `log-server` grpc interfaces, enabling restatectl to depend on log-server grpc client only without pulling everything else.
- Note that next PR removes `memory-loglet` gating from non-bifrost crates.
- Moves `NodeSetChecker` and `logserver_candidate_filter` into `restate-types` to unblock removing log-server as a dependency of restatectl
- Removes memory-loglet gating except for bifrost
- Removes unnecessary dependencies from restatectl
- Removes stale `restatectl metadata` options that were not cleaned up after moving to metadata proxy

```
// intentionally empty
```