Integrate restate_core::cluster_state::ClusterState into the ClusterController
This PR changes how the ClusterController learns about alive/dead nodes. Now
it uses the restate_core::cluster_state::ClusterState to update the ObservedClusterState
which is used by the LogsController and the Scheduler.

As part of this commit, we hide the implementation details of the ObservedClusterState
to allow for easier evolutions of it. Moreover, it simplifies the error handling of the
LogsController by panicking if the loglet params cannot be parsed because it is a
situation from which the Restate server cannot recover until the metadata gets fixed.