Introduce WorkerState
This commit introduces the WorkerState which is stored in a node's NodeConfig.
The WorkerState (together with the role) describes whether a node can run partitions.
Moreover, it allows to control the lifecycle of the worker role. By setting the state
to WorkerState::Draining, this node will no longer be added to new replica sets and
existing ones will eventually be moved to other nodes if possible. This allows to
gracefully drain a given node before it gets removed from the cluster.

This fixes #3260.