Only transition node's MetadataServerState from Member to Standby if node was member before
Since we are currently using the MetadataServerState::Member for nodes that have successfully
joined a metadata server cluster as well as for newly started Restate servers that are supposed
to join the metadata cluster, we can't unconditionally set the MetadataServerState to Standby
for nodes that are not part of the metadata cluster configuration. Instead, we now only do it
if the node has been part of the previous configuration.

This fixes #3275.