Run SequencerAppender tasks in default runtime
We want to ensure that SequencerAppender tasks complete to update the sequencer state.
Therefore, we run the tasks in the default runtime to avoid not completing them if they
run as part of the partition processor runtime.

Before, it could happen that a SequencerAppender did not finish (because the corresponding
partition processor runtime was shut down) and did not update the sequencer shared state
once the store succeeded. This would prevent subsequent append operations to hang for as
long as the periodic tail check needs to learn about the local tails of the log servers
(~15s).