Fix first_attempt detection when updating the Storage- and WorkerState
Because the inner closure moved references and the first_attempt variable is bool,
it was copied into the inner closure. Therefore, repeated attempts to update the
Storage/WorkerState would always see first_attempt == true. As a consequence the
server would panic when retrying and still thinking that it was the first attempt
to update these values.