Metadata sync messages to use normal connection abstractions
Since the metadata syncing is now properly non-blocking, it's best to use normal connection abstraction for this. This closes the last gap in graceful connection drains as previously it was possible for a connection to hang at drain time when if it waited for the metadata update rpc response for too long.