BalancedSpreadSelector initial implementation
This is the first iteration on a spread selector that's currently tuned for choosing replica-sets for partitions. The plan is to adapt it and clean it up so it's also used in bifrost's replicated-loglet spread selection. The other use case will benefit from some performance optimizations like sharing the topology and generalizing it so it overlays the entire nodes configuration instead of a filtered view.

The balanced heap strategy attempts to spread selection across mandatory scopes (region, zone) before backfilling with other nodes. It satisfies the domain-aware replication property requirements and supports consistent regeneration.

```
// intentionally empty.
```