Fix change from rebase
Feedback

Rework how we store the events. Now events are stored alike other journal entries, with the difference that the event protobuf definition is defined one for all the protocol versions.

First feedback

Add just docker-local-fedora as temporary solution to iterate fast locally on e2e tests.

Add new integration test run with journal retention and journal events enabled

Rebase changes

Update cargo toml

[invoker] Add experimental feature for invoker events propagation.

We need this because of the bug solved in the previous commit.

[invoker] Fix bug that breaks replaying an invocation when the journal contains Events. Without this fix, the SDK will expect more entries than the one we effectively send through.

[invoker] Implement propagation of transient error events.

This happens only when we're sure that the selected protocol is v4+.

[pp state machine] Implement deduplication of the TransientError event. This makes sure we don't store too many duplicated errors.

Changes related to the previous commit

[types] Refactor the Event data type.
With this change events have a more high level representation when materialized to the `Entry` data mode, but they still retain the same JSON and Protobuf representation at `RawEntry` level.
This change is storage backward compatible, but will result in a different display of events when using `sys_journal.entry_json` (which is fine, as previous runtime version didn't use events at all)