Replace ControlProcessors message with PartitionReplicaSetStates in the PPM
The PPM now uses the PartitionRepliacSetStates to decide which partition processors to
start and stop. Only which processor should become the leader is now decided by
the scheduler which is communicated via the ControlProcessor message.

Relying on the PartitionReplicaSetStates also helped simplifying the ProcessorState enum
which no longer needs to remember whether to restart or not because it is defined by the
PartitionReplicaSetStates. Morever, we assume that the scheduler will retry selecting a
partition processor as leader if there is a race condition between telling that a pp should
be the leader and the arrival of the PartitionReplicaSetStates.