[Bifrost] Auto-improvement of logs on preferred writer nodes
Bifrost will not gradually auto-reconfigure logs based on its knowledge of preferred writers of this log running on a node. The preferred writer in this context is the partition processor leader.

The intent is implied from the supplied error recovery strategy that the appender gets on creation.

Configuration: A new _hidden_ configuration `bifrost.disable_auto_improvement` is set to `false` by default that can disable this feature.

```
// intentionally empty
```