[Gossip] Nodes delay their own suspect->alive transition slightly
This delay (600ms) addresses the synchronized startup scenario where all nodes will transition into alive and compete on admin leadership for a brief period. It does not eliminate the problem completely of course but it has experimentally proven to improve the situation significantly.

This also updates the default gossip duration to be 100ms. Future adjustments are subject to lab-bench testing.