Replace bounded partition processor control message channel with watch
We currently cannot tolerate dropping PartitionProcessorControlMessages because
the upper layers assume that the message will be seen and acted upon by the
PartitionProcessor. To solve this problem, this commit replaces the bounded
channel with a watch since the partition processor should only be interested in
the last target leader state that the PPM instructed it to become.