Reject write requests early in the RaftMetadataServer
This commit tries to reject impossible write requests early before replicating
them through Raft in the replicated metadata store. It does that by comparing
the version of the stored value when receiving the request with the requests
precondition. If the stored version is already larger than the preconditioned
version, then we concluded that the write cannot succeed (assuming monotonically
increasing versions).

Additionally, this commit changes the RaftMetadataServer's event loop to be a biased
select statement. This gives precendence to processing raft messages over external
requests. Otherwise, there is the risk that external requests affect the Raft
state machine negatively.