Coalesce log updates via the LogChainExtender
In order to avoid thrashing when there are multiple appenders that want to extend
their log chains and therefore try to update the Logs configuration at the same
time, this commit introduces the LogChainExtender. The LogChainExtender is a background
task that tries to coalesce log chain updates into a single Logs write.