Update PartitionRouting to rely in PartitionReplicaSetStates and ClusterState
This commit updates the PartitionRouting to use the PartitionReplicaSetStates and ClusterState
to find the current leader. Additionally, it stops the Scheduler from updating the PartitionTable's
PartitionPlacement on every leader change.

The PartitionPlacement in the PartitionTable is replaced by the PartitionReplicaSetStates
and the ClusterState. Therefore, we can remove the PartitionPlacement which was an optional
field in the PartitionTable anyway.