Split ClusterController reponsibilities into separate tasks
This commit introduces the TrimLogsTask which encapsulates the trim logs logic
that previously ran as part of the ClusterController's main event loop. Now, it
runs in a separate task to avoid blocking the main event loop.

Likewise, this commit introduces the SchedulerTask which runs the ClusterController's
Scheduler if it is the leader. The Scheduler activity is now completely encapsulated
in this task. Consequently, the Scheduler will no longer block the ClusterController's
main event loop.