Set the initial current partition configuration version to be equal to the EpochMetadata's version (#3386)
* Set the initial current partition configuration version to be equal to the EpochMetadata's version

By setting the initial current partition configuration version to be equal to the EpochMetadata's
version we ensure that clobbering the EpochMetadata (e.g. during a rolling upgrade) won't cause
a diverging view on the cluster's ReplicaSetState. W/o this fix, it was possible that the system
was reusing versions for different partition configurations which would have cause the system
to end up with a divergent view on which node is running which partition processors.

* Reset Scheduler's PartitionState if current partition configuration is invalid

In order to better handle the situation where old versions clobber the EpochMetadata,
the Scheduler's PartitionState will be reset if the current partition configuration is
invalid. This will ensure that on the next event/tick, the scheduler will try to find
a valid current partition configuration.