[StateMachine] Semantic version barrier support
Version barriers is a protection mechanism for future versions of restate. The barrier allows the log to be used to guard against old/unsupported restate versions from processing messages at or after the barrier is observed.

The version barrier is stored in the partition store as a new FSM variable and defaults to 0.0.0-unknown which makes any version of restate compatible by default. Note that the barrier reason is _not_ stored in the partition store, we only use the `human_reason` to log the error during replay
but the reason will not be persisted in the partition store. So a follower that fetches a snapshot of the partition with higher min version will not show the reason of the barrier.