Print version notice on restate-server releases
With this change, restate-server will run a check in the background to fetch the latest version from github releases  and prints a nice eye-candy notice
to encourage users to update to the latest version.

This doesn't affect startup time, and fails silently if github is down/unreachable for any reason.