[datafusion] Decouples the physical/logical partition scanners
This commit decouples the number of active scanners for a given
datafusion query. Previously we would have a scanner (rocksdb or rpc)
for every partition store at the same time, with this commit we bundle
partitions to groups such that the total number of groups will be
`query-parallelism` (or in datafusion world `target_partitions`).
Each group will scan it's partition stores in order of its
partition-ids.
With this we can control the number of concurrent physical scanners.