Enable http-ingress role by default
This takes a slightly different approach than #3140 as it'll still start ingress even if the role `http-ingress` is not set. A notice message will be printed for nodes running with `worker` and without `http-ingress` to ease users into the change. In v1.5, http-ingress will not automatically start with worker and it'll require the role to be set explicitly.

In v1.4.0 the role `http-ingress` is added to the default set of roles so users starting with the default configuration will not see the notice.