[fix] Fix invocation timestamps non-determinisim (#3471)
Invocation creation time is now set to rely on Bifrost's `created_at` (Et al.) field(s) as a stable source of truth.
This fixes a bug where invocations can appear to be created at a later time than they actually when repaying the log.

Additionally, this commit also puts back the use of Arc for envelopes that was removed in #3361. This is to make sure
we don't forget to reduce the number of clones we are doing if the record is cached, and in downstream.

Fixes #3483

```
// intentionally empty
```

***
Fix the rest of invocation status timestamps