Explicitly set journal append_time based on Bifrost's record creation time
Currently, journal entries derived from signals use now as the append time for the created
RawEntryHeader. This is unlike other journal entries which set create the RawEntryHeader before
writing the journal entry to Bifrost. Due to this, the append time of signals is set at
application time which can be different than the other journal entries' append time. To ensure
that the append time is monotonoically increasing, this commit overwrites the RawEntry's append
time with Bifrost's record creation time.

This fixes #3487.