Prepare not sending RawEntryHeaders through Bifrost
Right now, the RawEntryHeaders are only used to attach meta information to
a RawEntryInner that originates when applying the Bifrost record. To save
bytes that need to be replicated through Bifrost, this commit prepares the
restate_invoker_api::EffectKind::JournalEntryV2 to only contain the
RawEntryInner. This is achieved by making the entry field optional and
adding an optional field raw_entry. The former field will no longer be set
used in v1.5.0 when the latter field will be populated.

Rename RawEntry into StoredRawEntry and RawEntryInner into RawEntry

Move StoredRawEntry and StoredRawEntryHeader to crates/types/src/storage.rs