[Fix] Fix query-engine deadlock when contending on lo-pri thread pool
This PR fixes a deadlock that can occur when remoq query server's network service loop gets blocked trying to open a new iterator when the low-priority thread pool is exhausted by currently open scanners. The deadlock happens because the existing iterators cannot make progress or expire because they are managed by the top-level `tokio::select!`.

The fix addresses this issue by moving into a concurrent design (remote scanners as independent tasks). Each scanner manages its own expiration and lifecycle.


Additionally, the PR fixes another issue where cancellation of in-flight query didn't remove the scanner from the list of active scanners.
This means that Ctrl+C of a long `restate sql` would block the low-priority threads until those scanners expire (60s by default).

The cancellation signal from the admin node is now propagated correctly to the remote scanner's node if it's possible to do so.
Additionally, the remote scanner will now monitor if the peer node is dead and will auto-remove the remote scanner (aka. the iterator) when the peer is marked as dead in gossip. This addresses
the scenario where the admin node is killed/shutdown while the query is still running, causing the remote scanners on peers to linger until the expiration to kick in.

```
// intentionally empty
```